+++ NEWS.md
@@ -1,2 +1,8 @@
+# CAISER 0.2.2
+* Added new example and use case to `calc_nreps2()`
+
+# CAISEr 0.2.1
+* Minor fixes, particularly in printing function.
+
 # CAISEr 0.2.0
-* Initial release on CRAN
+* Initial release on CRAN.